requirements.txt
httpx>=0.27.0
pandas>=2.2.0
pyarrow>=16.0.0
numpy>=1.26.0
apscheduler>=3.10.4
loguru>=0.7.2
pydantic-settings>=2.2.0
pytz>=2024.1
websockets>=12.0
# dashboard
streamlit>=1.33.0
streamlit-autorefresh>=1.0.1
plotly>=5.20.0
